Split a few function in a shared library
For now, only the functions that I do need for another feature
were moved, but later I want to move all of them (those that can
be moved), and with enough time and courage, maybe start to
split the project on its own.
